+++ tapis-notificationslib/src/main/java/edu/utexas/tacc/tapis/notifications/gen/jooq/tables/NotificationsTests.java
@@ -14,14 +14,18 @@ import edu.utexas.tacc.tapis.notifications.gen.jooq.tables.records.Notifications
 import java.time.LocalDateTime;
 import java.util.Arrays;
 import java.util.List;
+import java.util.function.Function;
 
 import org.jooq.Field;
 import org.jooq.ForeignKey;
+import org.jooq.Function10;
 import org.jooq.Identity;
 import org.jooq.Name;
 import org.jooq.Record;
-import org.jooq.Row9;
+import org.jooq.Records;
+import org.jooq.Row10;
 import org.jooq.Schema;
+import org.jooq.SelectField;
 import org.jooq.Table;
 import org.jooq.TableField;
 import org.jooq.TableOptions;
@@ -97,6 +101,11 @@ public class NotificationsTests extends TableImpl<NotificationsTestsRecord> {
      */
     public final TableField<NotificationsTestsRecord, LocalDateTime> UPDATED = createField(DSL.name("updated"), SQLDataType.LOCALDATETIME(6).nullable(false).defaultValue(DSL.field("timezone('utc'::text, now())", SQLDataType.LOCALDATETIME)), this, "");
 
+    /**
+     * The column <code>tapis_ntf.notifications_tests.start_count</code>.
+     */
+    public final TableField<NotificationsTestsRecord, Integer> START_COUNT = createField(DSL.name("start_count"), SQLDataType.INTEGER.nullable(false).defaultValue(DSL.field("0", SQLDataType.INTEGER)), this, "");
+
     private NotificationsTests(Name alias, Table<NotificationsTestsRecord> aliased) {
         this(alias, aliased, null);
     }
@@ -106,14 +115,16 @@ public class NotificationsTests extends TableImpl<NotificationsTestsRecord> {
     }
 
     /**
-     * Create an aliased <code>tapis_ntf.notifications_tests</code> table reference
+     * Create an aliased <code>tapis_ntf.notifications_tests</code> table
+     * reference
      */
     public NotificationsTests(String alias) {
         this(DSL.name(alias), NOTIFICATIONS_TESTS);
     }
 
     /**
-     * Create an aliased <code>tapis_ntf.notifications_tests</code> table reference
+     * Create an aliased <code>tapis_ntf.notifications_tests</code> table
+     * reference
      */
     public NotificationsTests(Name alias) {
         this(alias, NOTIFICATIONS_TESTS);
@@ -132,7 +143,7 @@ public class NotificationsTests extends TableImpl<NotificationsTestsRecord> {
 
     @Override
     public Schema getSchema() {
-        return TapisNtf.TAPIS_NTF;
+        return aliased() ? null : TapisNtf.TAPIS_NTF;
     }
 
     @Override
@@ -146,17 +157,21 @@ public class NotificationsTests extends TableImpl<NotificationsTestsRecord> {
     }
 
     @Override
-    public List<UniqueKey<NotificationsTestsRecord>> getKeys() {
-        return Arrays.<UniqueKey<NotificationsTestsRecord>>asList(Keys.NOTIFICATIONS_TESTS_PKEY, Keys.NOTIFICATIONS_TESTS_TENANT_OWNER_SUBSCR_NAME_KEY);
+    public List<UniqueKey<NotificationsTestsRecord>> getUniqueKeys() {
+        return Arrays.asList(Keys.NOTIFICATIONS_TESTS_TENANT_OWNER_SUBSCR_NAME_KEY);
     }
 
     @Override
     public List<ForeignKey<NotificationsTestsRecord, ?>> getReferences() {
-        return Arrays.<ForeignKey<NotificationsTestsRecord, ?>>asList(Keys.NOTIFICATIONS_TESTS__NOTIFICATIONS_TESTS_SUBSCR_SEQ_ID_FKEY);
+        return Arrays.asList(Keys.NOTIFICATIONS_TESTS__NOTIFICATIONS_TESTS_SUBSCR_SEQ_ID_FKEY);
     }
 
     private transient Subscriptions _subscriptions;
 
+    /**
+     * Get the implicit join path to the <code>tapis_ntf.subscriptions</code>
+     * table.
+     */
     public Subscriptions subscriptions() {
         if (_subscriptions == null)
             _subscriptions = new Subscriptions(this, Keys.NOTIFICATIONS_TESTS__NOTIFICATIONS_TESTS_SUBSCR_SEQ_ID_FKEY);
@@ -172,6 +187,11 @@ public class NotificationsTests extends TableImpl<NotificationsTestsRecord> {
     @Override
     public NotificationsTests as(Name alias) {
         return new NotificationsTests(alias, this);
+    }
+
+    @Override
+    public NotificationsTests as(Table<?> alias) {
+        return new NotificationsTests(alias.getQualifiedName(), this);
     }
 
     /**
@@ -190,12 +210,35 @@ public class NotificationsTests extends TableImpl<NotificationsTestsRecord> {
         return new NotificationsTests(name, null);
     }
 
+    /**
+     * Rename this table
+     */
+    @Override
+    public NotificationsTests rename(Table<?> name) {
+        return new NotificationsTests(name.getQualifiedName(), null);
+    }
+
     // -------------------------------------------------------------------------
-    // Row9 type methods
+    // Row10 type methods
     // -------------------------------------------------------------------------
 
     @Override
-    public Row9<Integer, Integer, String, String, String, Integer, JsonElement, LocalDateTime, LocalDateTime> fieldsRow() {
-        return (Row9) super.fieldsRow();
+    public Row10<Integer, Integer, String, String, String, Integer, JsonElement, LocalDateTime, LocalDateTime, Integer> fieldsRow() {
+        return (Row10) super.fieldsRow();
+    }
+
+    /**
+     * Convenience mapping calling {@link SelectField#convertFrom(Function)}.
+     */
+    public <U> SelectField<U> mapping(Function10<? super Integer, ? super Integer, ? super String, ? super String, ? super String, ? super Integer, ? super JsonElement, ? super LocalDateTime, ? super LocalDateTime, ? super Integer, ? extends U> from) {
+        return convertFrom(Records.mapping(from));
+    }
+
+    /**
+     * Convenience mapping calling {@link SelectField#convertFrom(Class,
+     * Function)}.
+     */
+    public <U> SelectField<U> mapping(Class<U> toType, Function10<? super Integer, ? super Integer, ? super String, ? super String, ? super String, ? super Integer, ? super JsonElement, ? super LocalDateTime, ? super LocalDateTime, ? super Integer, ? extends U> from) {
+        return convertFrom(toType, Records.mapping(from));
     }
 }