+++ tapis-notificationslib/src/main/java/edu/utexas/tacc/tapis/notifications/model/Event.java
@@ -1,8 +1,10 @@
 package edu.utexas.tacc.tapis.notifications.model;
 
+import edu.utexas.tacc.tapis.shared.utils.TapisUtils;
 import org.apache.commons.lang3.StringUtils;
 
-import java.net.URI;
+import java.time.Instant;
+import java.time.ZoneOffset;
 import java.util.UUID;
 import java.util.regex.Pattern;
 
@@ -11,6 +13,14 @@ import java.util.regex.Pattern;
  * Subscriptions are used to request that notifications be sent when Events are received.
  * Receipt of an event can result in 0 or more notifications.
  * It is possible for an event be received and then effectively discarded because there are no subscriptions for it.
+ *
+ * SeriesId:
+ *   Each series is intended to sequentially track events of various types coming from a
+ *   given tenant, source and subject. So for each tenant, source and subject the seriesId is considered unique.
+ *   For example, the Jobs service (the source) sends out events with the jobUuid as the subject and
+ *   sets the seriesId to the jobUuid. That way a subscription can be created to follow (in order) all
+ *   events of various types related to the job.
+ *   Examples of event types defined in the Jobs service: JOB_NEW_STATUS, JOB_ERROR_MESSAGE, JOB_USER_EVENT
  *
  * Based on the CloudEvent specification.
  * Represents an event associated with an occurrence. An Occurrence may produce multiple events.
@@ -33,10 +43,13 @@ public final class Event
 
   // Default values
   public static final boolean DEFAULT_DELETE_SUBSCRIPTIONS_MATCHING_SUBJECT = false;
+  public static final boolean DEFAULT_END_SERIES = false;
+  public static final long DEFAULT_SERIES_SEQ_COUNT = -1L;
 
   // Valid pattern for event type, must be 3 sections separated by a '.'
   // First section must contain a series of lower case letters and may not be empty
-  // Second and third sections must start alphabetic, contain only alphanumeric or 3 special characters: - _ ~ and may not be empty
+  // Second and third sections must start alphabetic, contain only alphanumeric
+  //   or 3 special characters: - _ ~ and may not be empty
   private static final Pattern EVENT_TYPE_PATTERN =
           Pattern.compile("^[a-z]+\\.[a-zA-Z]([a-zA-Z0-9]|[-_~])*\\.[a-zA-Z]([a-zA-Z0-9]|[-_~])*$");
 
@@ -50,10 +63,15 @@ public final class Event
   private final String subject; // Subject of event in context of event producer.
   private final String data; // Data associated with the event.
   private final String seriesId; // Optional Id for grouping events from same source.
+  private final long seriesSeqCount; // Sequence counter associated with seriesId for ordering of events from same source.
   private final String timestamp; // Timestamp of when the occurrence happened. RFC 3339 (ISO 8601)
-  private final boolean deleteSubscriptionsMatchingSubject;
+  private final boolean deleteSubscriptionsMatchingSubject;  // Indicates all subscriptions associated with subject
+                                                             //   should be removed after deliveries are complete.
+  private final boolean endSeries;  // Indicates that this is the last event in a series.
+                                    // Tracking data will be removed after deliveries are complete.
   private final String tenant; // Tenant associated with the event
   private final String user; // User associated with the event
+  private final Instant received; // Timestamp of when the event was received by the Notifications service.
   private final UUID uuid;
   // Mark as transient so Gson will not include it.
   private transient String type1; // Field 1 of type (service name)
@@ -63,18 +81,22 @@ public final class Event
   /* ********************************************************************** */
   /*                           Constructors                                 */
   /* ********************************************************************** */
-  public Event(String source1, String type1, String subject1, String data1, String seriesId1, String timestamp1,
-               boolean deleteSubscriptionsMatchingSubject1, String tenant1, String user1, UUID uuid1)
+  public Event(String source1, String type1, String subject1, String data1, String seriesId1,
+               long seriesSeqCount1, String timestamp1, boolean deleteSubscriptionsMatchingSubject1,
+               boolean endSeries1, String tenant1, String user1, Instant received1, UUID uuid1)
   {
     source = source1;
     type = type1;
     subject = subject1;
     data = data1;
     seriesId = seriesId1;
+    seriesSeqCount = seriesSeqCount1;
     timestamp = timestamp1;
     deleteSubscriptionsMatchingSubject = deleteSubscriptionsMatchingSubject1;
+    endSeries = endSeries1;
     tenant = tenant1;
     user = user1;
+    received = (received1 != null) ? received1 : TapisUtils.getUTCTimeNow().toInstant(ZoneOffset.UTC);
     uuid = uuid1;
     setTypeFields();
   }
@@ -87,11 +109,14 @@ public final class Event
   public String getSubject() { return subject; }
   public String getData() { return data; }
   public String getSeriesId() { return seriesId; }
+  public long getSeriesSeqCount() { return seriesSeqCount; }
   public String getTimestamp() { return timestamp; }
   public boolean getDeleteSubscriptionsMatchingSubject() { return deleteSubscriptionsMatchingSubject; }
+  public boolean getEndSeries() { return endSeries; }
 
   public String getTenant() { return tenant; }
   public String getUser() { return user; }
+  public Instant getReceived() { return received; }
   public UUID getUuid() { return uuid; }
   public String getType1() { return type1; }
   public String getType2() { return type2; }
@@ -110,8 +135,13 @@ public final class Event
   @Override
   public String toString()
   {
-    String msg = "Source: %s Type: %s Subject: %s Data: %s SeriesId: %s Timestamp: %s UUID: %s";
-    return msg.formatted(source, type, subject, data, seriesId, timestamp, uuid);
+    String msg =
+"""
+ Tenant: %s Source: %s Type: %s Subject: %s Data: %s SeriesId: %s SeriesSeqCount: %d Timestamp: %s
+ deleteSubscriptionsMatchingSubject: %b endSeries: %b user: %s received: %s UUID: %s";
+""";
+    return msg.formatted(tenant, source, type, subject, data, seriesId, seriesSeqCount, timestamp,
+                         deleteSubscriptionsMatchingSubject, endSeries, user, received, uuid);
   }
 
   /*