+++ tapis-notificationslib/src/test/java/edu/utexas/tacc/tapis/notifications/IntegrationUtils.java
@@ -13,6 +13,7 @@ import edu.utexas.tacc.tapis.shared.utils.TapisGsonUtils;
 import edu.utexas.tacc.tapis.shared.utils.TapisUtils;
 
 import java.time.Instant;
+import java.time.ZoneOffset;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -133,13 +134,16 @@ public final class IntegrationUtils
   public static final String eventType1 = "jobs.job.complete";
   public static final String eventSubject1 = "640ad5a8-1a6e-4189-a334-c4c7226fb9ba-007";
   public static final String seriesId1 = "111a2228-1a6e-4189-a334-c4c722666666-007";
+  public static final int seriesSeqCount1 = 1;
   public static final String eventDataNull = null;
   public static final String eventTime = TapisUtils.getUTCTimeNow().toString();
   public static final boolean eventDeleteSubscriptionsMatchingSubjectFalse = false;
+  public static final boolean eventEndSeriesFalse = false;
 
-  public static final Event event1 = new Event(eventSource1, eventType1, eventSubject1, eventDataNull, seriesId1,
-                                               eventTime, eventDeleteSubscriptionsMatchingSubjectFalse, tenantName,
-                                               testUser1, UUID.randomUUID());
+  public static final Event event1 = new Event(eventSource1, eventType1, eventSubject1, eventDataNull, seriesId1, seriesSeqCount1,
+                                               eventTime, eventDeleteSubscriptionsMatchingSubjectFalse, eventEndSeriesFalse,
+                                               tenantName, testUser1,
+                                               TapisUtils.getUTCTimeNow().toInstant(ZoneOffset.UTC), UUID.randomUUID());
 
   /**
    * Create an array of Subscription objects in memory